use tristate for can apply change event
