+++ src/main/java/com/github/thedeathlycow/thermoo/impl/LivingEntityEnvironmentTickImpl.java
@@ -3,14 +3,18 @@ package com.github.thedeathlycow.thermoo.impl;
 import com.github.thedeathlycow.thermoo.api.temperature.EnvironmentController;
 import com.github.thedeathlycow.thermoo.api.temperature.EnvironmentManager;
 import com.github.thedeathlycow.thermoo.api.temperature.HeatingModes;
+import com.github.thedeathlycow.thermoo.api.temperature.event.PlayerEnvironmentEvents;
 import com.github.thedeathlycow.thermoo.impl.component.ThermooComponents;
+import net.fabricmc.fabric.api.util.TriState;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
 public class LivingEntityEnvironmentTickImpl {
 
 
-    public static void tick(LivingEntity entity, int lastTickTemperature) {
+    public static void tick(LivingEntity entity) {
         World world = entity.getWorld();
 
         if (world.isClient() || entity.isSpectator() || entity.isDead() || entity.isRemoved()) {
@@ -46,6 +50,30 @@ public class LivingEntityEnvironmentTickImpl {
 
         if (isSyncTick || ThermooComponents.WETNESS.get(entity).isDirty()) {
             ThermooComponents.WETNESS.sync(entity);
+        }
+    }
+
+    public static void tickPlayer(PlayerEntity player) {
+        World world = player.getWorld();
+
+        if (world.isClient || player.isSpectator()) {
+            return;
+        }
+
+        BlockPos pos = player.getBlockPos();
+        var controller = EnvironmentManager.INSTANCE.getController();
+        int temperatureChange = controller.getLocalTemperatureChange(world, pos);
+
+        if (temperatureChange != 0) {
+
+            TriState canApplyChange = PlayerEnvironmentEvents.CAN_APPLY_PASSIVE_TEMPERATURE_CHANGE
+                    .invoker()
+                    .canApplyChange(temperatureChange, player);
+
+            if (canApplyChange == TriState.TRUE) {
+                temperatureChange = controller.getEnvironmentTemperatureForPlayer(player, temperatureChange);
+                player.thermoo$addTemperature(temperatureChange, HeatingModes.PASSIVE);
+            }
         }
     }
