+++ src/main/java/com/github/thedeathlycow/thermoo/mixin/common/PlayerTemperatureTickMixin.java
@@ -1,14 +1,8 @@
 package com.github.thedeathlycow.thermoo.mixin.common;
 
-import com.github.thedeathlycow.thermoo.api.temperature.EnvironmentManager;
-import com.github.thedeathlycow.thermoo.api.temperature.HeatingModes;
-import com.github.thedeathlycow.thermoo.api.temperature.event.PlayerEnvironmentEvents;
-import net.minecraft.entity.Entity;
+import com.github.thedeathlycow.thermoo.impl.LivingEntityEnvironmentTickImpl;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -27,26 +21,7 @@ public abstract class PlayerTemperatureTickMixin {
     )
     private void onTick(CallbackInfo ci) {
         PlayerEntity player = (PlayerEntity) (Object) this;
-        World world = player.getWorld();
-
-        if (world.isClient || player.isSpectator()) {
-            return;
-        }
-
-        BlockPos pos = player.getBlockPos();
-        var controller = EnvironmentManager.INSTANCE.getController();
-        int temperatureChange = controller.getLocalTemperatureChange(world, pos);
-
-        boolean canApplyChange = temperatureChange != 0 &&
-                PlayerEnvironmentEvents.CAN_APPLY_PASSIVE_TEMPERATURE_CHANGE
-                        .invoker()
-                        .canApplyChange(temperatureChange, player);
-
-        if (canApplyChange) {
-            temperatureChange = controller.getEnvironmentTemperatureForPlayer(player, temperatureChange);
-            player.thermoo$addTemperature(temperatureChange, HeatingModes.PASSIVE);
-        }
-
+        LivingEntityEnvironmentTickImpl.tickPlayer(player);
     }
 
 }
